refactor: create useContent custom hook
useContent custom hook을 만들고 관련 내용을 useInterview에서 수정.
FooterR에서 관련 내용 수정.
